backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
prometheus-client==0.19.0

# Image processing and vectorization
scikit-image==0.22.0
opencv-python-headless==4.8.1.78
potrace==0.2.0

# Background tasks
celery==5.3.4
kombu==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Hashing for cache keys
xxhash==3.4.1
